+++ payment_api/src/providers/schemas.py
@@ -1,21 +1,22 @@
 from src.schemas.base import AbstractSchema
 
 
-class ProviderPayment(AbstractSchema):
+class ProviderBase(AbstractSchema):
     amount: int
     customer: str
     currency: str = "RUB"
+
+
+class ProviderPayment(ProviderBase):
     setup_future_usage: str = 'off_session'
-    confirm = False
+    confirm: bool = False
+
+
+class ProviderPaymentCancel(ProviderBase):
+    payment: str
 
 
 class ProviderPaymentResult(AbstractSchema):
     id: str
     client_secret: str
 
-
-class ProviderPaymentCancel(AbstractSchema):
-    customer: str
-    amount: int
-    payment: str
-    currency: str = "RUB"
