+++ tests/integration/conftest.py
@@ -8,8 +8,10 @@ import pytest
 import jwt
 from multidict import CIMultiDictProxy
 
-from .settings import (ALGORITHM, SECRET_KEY, USER_ID, SUBSCRIPTION_API_URL, SUBSCRIPTION_API_VERSION, PAYMENTS_API_URL,
-                       PAYMENTS_API_VERSION)
+from settings import (
+    ALGORITHM, SECRET_KEY, USER_ID, SUBSCRIPTION_API_URL,
+    SUBSCRIPTION_API_VERSION, PAYMENTS_API_URL, PAYMENTS_API_VERSION
+)
 
 
 @dataclass
